+++ doc/changelog.rst
@@ -15,6 +15,7 @@ Unreleased (fork by JustAnotherArchivist)
 * Added: URL priorisation through `--priority-*` options and a `get_priority` hook.
 * Changed: `wpull.pipeline.pipeline.ItemSource` now deals in item generation coroutines rather than items directly.
 * Changed: wpull no longer flattens consecutive slashes in the path component of URLs, in line with RFC 3986, related standards, and behaviour of other software (wget and browsers).
+* Changed: Backslashes in URL paths are now treated like forward slashes in accordance with the URL Standard and behaviour of browsers.
 
 
 Backwards incompatibility